+++ src/ts/UIFactory.ts
@@ -15,6 +15,7 @@ import { PlaybackTimeLabel, PlaybackTimeLabelMode } from './components/labels/Pl
 import { SeekBar } from './components/seekbar/SeekBar';
 import { SeekBarLabel } from './components/seekbar/SeekBarLabel';
 import { PlaybackToggleButton } from './components/buttons/PlaybackToggleButton';
+import { QuickSeekButton } from './components/buttons/QuickSeekButton';
 import { VolumeToggleButton } from './components/buttons/VolumeToggleButton';
 import { VolumeSlider } from './components/seekbar/VolumeSlider';
 import { Spacer } from './components/Spacer';
@@ -187,7 +188,7 @@ export namespace UIFactory {
       player,
       [
         {
-          ui: () => UIFactory.defaultLayouts.castReceiver(config),
+          ui: UIFactory.defaultLayouts.castReceiver,
           identifier: UIVariantIdentifier.castReceiver,
         },
       ],
@@ -286,7 +287,6 @@ export namespace UIFactory {
       const settingsPanel = buildDefaultSettingsPanel(
         subtitleOverlay,
         undefined,
-        config.ecoMode === true,
         config.showPersistentPreferencesToggle === true && config.disableStorageApi !== true,
       );
       const controlBar = new ControlBar({
@@ -307,7 +307,9 @@ export namespace UIFactory {
           }),
           new Container({
             components: [
+              new QuickSeekButton({ seekSeconds: -10 }),
               new PlaybackToggleButton(),
+              new QuickSeekButton({ seekSeconds: 10 }),
               new VolumeToggleButton(),
               new VolumeSlider(),
               new Spacer(),
@@ -323,12 +325,6 @@ export namespace UIFactory {
         ],
       });
 
-      const conditionalComponents = [
-        config.includeWatermark ? new Watermark() : null,
-        ...(playerInsightsPanel ? [playerInsightsPanel] : []),
-        ...(playerContextMenu ? [new DismissClickOverlay({ target: playerContextMenu }), playerContextMenu] : []),
-      ].filter(e => e);
-
       return new UIContainer({
         components: [
           subtitleOverlay,
@@ -338,7 +334,9 @@ export namespace UIFactory {
           controlBar,
           new TitleBar(),
           new RecommendationOverlay(),
-          ...conditionalComponents,
+          new Watermark(),
+          ...(playerInsightsPanel ? [playerInsightsPanel] : []),
+          ...(playerContextMenu ? [new DismissClickOverlay({ target: playerContextMenu }), playerContextMenu] : []),
           new DismissClickOverlay({ target: settingsPanel }),
           settingsPanel,
           new ErrorMessageOverlay(),
@@ -429,7 +427,9 @@ export namespace UIFactory {
           }),
           new Container({
             components: [
+              new QuickSeekButton({ seekSeconds: -10 }),
               new PlaybackToggleButton(),
+              new QuickSeekButton({ seekSeconds: 10 }),
               new VolumeToggleButton(),
               new VolumeSlider(),
               new Spacer(),
@@ -533,7 +533,7 @@ export namespace UIFactory {
       });
     }
 
-    export function castReceiver(config: UIConfig = {}): UIContainer {
+    export function castReceiver(): UIContainer {
       const controlBar = new ControlBar({
         components: [
           new Container({
@@ -553,8 +553,6 @@ export namespace UIFactory {
         ],
       });
 
-      const conditionalComponents = [config.includeWatermark ? new Watermark() : null].filter(e => e);
-
       return new CastUIContainer({
         components: [
           new SubtitleOverlay(),
@@ -562,7 +560,7 @@ export namespace UIFactory {
           new PlaybackToggleOverlay(),
           controlBar,
           new TitleBar({ keepHiddenWithoutMetadata: true }),
-          ...conditionalComponents,
+          new Watermark(),
           new ErrorMessageOverlay(),
         ],
         cssClasses: ['ui-cast-receiver'],
@@ -748,7 +746,6 @@ export namespace UIFactory {
   function buildDefaultSettingsPanel(
     subtitleOverlay: SubtitleOverlay,
     hideDelay: number | undefined = undefined,
-    enableEcoMode: boolean = false,
     showPersistentPreferencesToggle: boolean = false,
   ): SettingsPanel<SettingsPanelConfig> {
     const settingsPanelConfig: SettingsPanelConfig = {
@@ -785,16 +782,14 @@ export namespace UIFactory {
       }),
     ];
 
-    if (enableEcoMode) {
-      const ecoModeContainer = new EcoModeContainer();
+    const ecoModeContainer = new EcoModeContainer();
 
-      ecoModeContainer.setOnToggleCallback(() => {
-        // forces the browser to re-calculate the height of the settings panel when adding/removing elements
-        settingsPanel.getDomElement().css({ width: '', height: '' });
-      });
+    ecoModeContainer.setOnToggleCallback(() => {
+      // forces the browser to re-calculate the height of the settings panel when adding/removing elements
+      settingsPanel.getDomElement().css({ width: '', height: '' });
+    });
 
-      components.unshift(ecoModeContainer);
-    }
+    components.unshift(ecoModeContainer);
 
     const mainSettingsPanelPage = new SettingsPanelPage({
       components,